+++ Apps.Jira/DataSourceHandlers/AssigneeDataSourceHandler.cs
@@ -1,23 +1,17 @@
 ﻿using Apps.Jira.Dtos;
-using Blackbird.Applications.Sdk.Common;
-using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 using Blackbird.Applications.Sdk.Common.Invocation;
 using RestSharp;
 
 namespace Apps.Jira.DataSourceHandlers;
 
-public class AssigneeDataSourceHandler : BaseInvocable, IAsyncDataSourceHandler
+public class AssigneeDataSourceHandler : JiraInvocable, IAsyncDataSourceHandler
 {
-    private IEnumerable<AuthenticationCredentialsProvider> Creds =>
-        InvocationContext.AuthenticationCredentialsProviders;
-
     public AssigneeDataSourceHandler(InvocationContext invocationContext): base(invocationContext) { }
 
     public async Task<Dictionary<string, string>> GetDataAsync(DataSourceContext context,
         CancellationToken cancellationToken)
     {
-        var client = new JiraClient(Creds);
         var baseEndpoint = "/user/search?maxResults=20&query=";
 
         if (!string.IsNullOrWhiteSpace(context.SearchString))
@@ -29,8 +23,8 @@ public class AssigneeDataSourceHandler : BaseInvocable, IAsyncDataSourceHandler
         do
         {
             var endpoint = baseEndpoint + $"&startAt={startAt}";
-            var request = new JiraRequest(endpoint, Method.Get, Creds);
-            var response = await client.ExecuteWithHandling<IEnumerable<UserDto>>(request);
+            var request = new JiraRequest(endpoint, Method.Get);
+            var response = await Client.ExecuteWithHandling<IEnumerable<UserDto>>(request);
 
             if (!response.Any())
                 break;
