+++ Apps.Jira/DataSourceHandlers/IssueStatusDataSourceHandler.cs
@@ -1,18 +1,14 @@
 ﻿using Apps.Jira.Dtos;
 using Apps.Jira.Models.Identifiers;
 using Blackbird.Applications.Sdk.Common;
-using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 using Blackbird.Applications.Sdk.Common.Invocation;
 using RestSharp;
 
 namespace Apps.Jira.DataSourceHandlers;
 
-public class IssueStatusDataSourceHandler : BaseInvocable, IAsyncDataSourceHandler
+public class IssueStatusDataSourceHandler : JiraInvocable, IAsyncDataSourceHandler
 {
-    private IEnumerable<AuthenticationCredentialsProvider> Creds =>
-        InvocationContext.AuthenticationCredentialsProviders;
-
     private readonly ProjectIdentifier _projectIdentifier;
 
     public IssueStatusDataSourceHandler(InvocationContext invocationContext, 
@@ -27,10 +23,8 @@ public class IssueStatusDataSourceHandler : BaseInvocable, IAsyncDataSourceHandl
         if (_projectIdentifier.ProjectKey == null)
             throw new Exception("Please specify project key first.");
         
-        var client = new JiraClient(Creds);
-        var request = new JiraRequest($"/project/{_projectIdentifier.ProjectKey}/statuses", Method.Get,
-            Creds);
-        var response = await client.ExecuteWithHandling<IEnumerable<StatusesWrapper>>(request);
+        var request = new JiraRequest($"/project/{_projectIdentifier.ProjectKey}/statuses", Method.Get);
+        var response = await Client.ExecuteWithHandling<IEnumerable<StatusesWrapper>>(request);
         
         return response
             .SelectMany(statuses => statuses.Statuses)
