+++ Apps.Jira/DataSourceHandlers/PriorityDataSourceHandler.cs
@@ -7,19 +7,15 @@ using RestSharp;
 
 namespace Apps.Jira.DataSourceHandlers;
 
-public class PriorityDataSourceHandler : BaseInvocable, IAsyncDataSourceHandler
+public class PriorityDataSourceHandler : JiraInvocable, IAsyncDataSourceHandler
 {
-    private IEnumerable<AuthenticationCredentialsProvider> Creds =>
-        InvocationContext.AuthenticationCredentialsProviders;
-
     public PriorityDataSourceHandler(InvocationContext invocationContext): base(invocationContext) { }
 
     public async Task<Dictionary<string, string>> GetDataAsync(DataSourceContext context,
         CancellationToken cancellationToken)
     {
-        var client = new JiraClient(Creds);
-        var request = new JiraRequest("/priority", Method.Get, Creds);
-        var response = await client.ExecuteWithHandling<IEnumerable<PriorityDto>>(request);
+        var request = new JiraRequest("/priority", Method.Get);
+        var response = await Client.ExecuteWithHandling<IEnumerable<PriorityDto>>(request);
         return response.ToDictionary(p => p.Id, p => p.Name);
     }
 }
