+++ Apps.Jira/DataSourceHandlers/ProjectDataSourceHandler.cs
@@ -7,24 +7,20 @@ using RestSharp;
 
 namespace Apps.Jira.DataSourceHandlers;
 
-public class ProjectDataSourceHandler : BaseInvocable, IAsyncDataSourceHandler
+public class ProjectDataSourceHandler : JiraInvocable, IAsyncDataSourceHandler
 {
-    private IEnumerable<AuthenticationCredentialsProvider> Creds =>
-        InvocationContext.AuthenticationCredentialsProviders;
-
     public ProjectDataSourceHandler(InvocationContext invocationContext): base(invocationContext) { }
 
     public async Task<Dictionary<string, string>> GetDataAsync(DataSourceContext context,
         CancellationToken cancellationToken)
     {
-        var client = new JiraClient(Creds);
         var endpoint = "/project/search?maxResults=20";
 
         if (!string.IsNullOrWhiteSpace(context.SearchString))
             endpoint += $"&query={context.SearchString}";
         
-        var request = new JiraRequest(endpoint, Method.Get, Creds);
-        var response = await client.ExecuteWithHandling<ProjectWrapper>(request);
+        var request = new JiraRequest(endpoint, Method.Get);
+        var response = await Client.ExecuteWithHandling<ProjectWrapper>(request);
         return response.Values.ToDictionary(p => p.Key, p => p.Name);
     }
 }
