+++ Apps.Jira/JiraRequest.cs
@@ -1,16 +1,11 @@
-﻿using Blackbird.Applications.Sdk.Common.Authentication;
-using RestSharp;
+﻿using RestSharp;
 
 namespace Apps.Jira
 {
     public class JiraRequest : RestRequest
     {
-        public JiraRequest(string endpoint, Method method,
-            IEnumerable<AuthenticationCredentialsProvider> authenticationCredentialsProviders)
-            : base(endpoint, method)
+        public JiraRequest(string endpoint, Method method) : base(endpoint, method)
         {
-            this.AddHeader("Authorization", 
-                authenticationCredentialsProviders.First(p => p.KeyName == "Authorization").Value);
         }
     }
 }
